Updated the AVTAlgorithms to ask for the entire dataset when using data contracts.
We now explicity call every reader with a data selection of the entire dataset.
This allows us to load file formats which only support contract / streaming  based data
loading. This supports is merely the brute force approach. In the future a better way needs to be found that uses streaming.