Fixed use case where all time values are the same in VisIt.
Nek5000 and other readers can report the same time value for all time
steps, because they don't want to every file in the file series
to determine the actual time step values. ParaView is unable to replicate
that behavior.

So instead in this use case we simulate the time values with numbers
that range from 1 to N where N is the number of time steps.